New Main(broken)
This is the point where I cloned the git here however, this portion of the code removed the old functionality in summurizer.py to remove user input. I have not changed main.py to reflect these changes so it is currently broken.

Changed files/worked on files~

summarizer.py
main.py